1. PacketFactory 함수 수정  - char 배열 -> SendBuffer  - 추후 Memory Pool로 발전
2. Packet -> GameEvent Converter 구현

3. 임시 Login Packet으로 접속 테스트 완료

+++ Server/JobGraph Study/LF_Stack.h
@@ -83,7 +83,7 @@ public:
 			if (nullptr == temp) return -2;
 
 			NODE* next = temp->next;
-			int result = temp->value;
+			size_t result = temp->value;
 
 			if (temp != top) continue;
 			if (CAS(&top, temp, next)) return result;
